requirements.txt
streamlit>=1.31
pandas>=2.0
google-generativeai>=0.5.0
